+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add `validation_context` option to `Form::Builder` (#101)
+- Add `validators` helper to `FieldComponent` (#101)
+- Add `optional?` and `required?` helpers to `FieldComponent` (#101)
 - Add `label_text` helper (#103)
 - Add `field_id` helper, backported from Rails 7.0 (#104)
 - Add `weekday_select` helper (#105)